services/station_service/requirements.txt
grpcio
grpcio-tools
protobuf
SQLAlchemy
psycopg2-binary
geoalchemy2
geoalchemy2[psycopg2]

